+++ app/Http/Controllers/API/TICKETS/TicketController.php
@@ -23,30 +23,34 @@ class TicketController extends Controller
         $api_token = $request->header('Authorization');
         $user = UserController::verify_token($api_token);
         if ($user) {
-            //get all ticket
-            if ($user->hasPerm('read-ticket')) {
-                $tickets = Ticket::all();
-            } else {
-                $tickets = Ticket::all()->where('user_id', $user->id);
-            }
-            //get assigned tickets
-            $assigned_tickets = Assigned::all();
-            //foreach assigned tickets get ticket id
-            $assigned_tickets_array = [];
-            foreach ($assigned_tickets as $assigned_ticket) {
-                $assigned_tickets_array[] = $assigned_ticket->ticket_id;
-            }
-            foreach ($tickets as $ticket) {
-                //if ticket id is in assigned tickets array then set assigned to true and get user id
-                if (in_array($ticket->id, $assigned_tickets_array)) {
-                    $ticket->assigned = true;
-                    $ticket->assigned_user = $assigned_ticket->user_id;
+            if ($user->hasVerifiedEmail()) {
+                //get all ticket
+                if ($user->hasPerm('read-ticket')) {
+                    $tickets = Ticket::all();
                 } else {
-                    $ticket->assigned = false;
+                    $tickets = Ticket::all()->where('user_id', $user->id);
                 }
-            }
+                //get assigned tickets
+                $assigned_tickets = Assigned::all();
+                //foreach assigned tickets get ticket id
+                $assigned_tickets_array = [];
+                foreach ($assigned_tickets as $assigned_ticket) {
+                    $assigned_tickets_array[] = $assigned_ticket->ticket_id;
+                }
+                foreach ($tickets as $ticket) {
+                    //if ticket id is in assigned tickets array then set assigned to true and get user id
+                    if (in_array($ticket->id, $assigned_tickets_array)) {
+                        $ticket->assigned = true;
+                        $ticket->assigned_user = $assigned_ticket->user_id;
+                    } else {
+                        $ticket->assigned = false;
+                    }
+                }
 
-            return response()->json($tickets, 200);
+                return response()->json($tickets, 200);
+            } else {
+                return response()->json(['error' => 'Verify email address'], 403);
+            }
         } else {
             return response()->json(['error' => 'Verfiy api token'], 403);
         }
@@ -65,20 +69,24 @@ class TicketController extends Controller
         $user = UserController::verify_token($api_token);
         //create ticket if user
         if ($user) {
-            $ticket = new Ticket();
-            $ticket->user_id = $user->id;
-            $ticket->title = $request->title;
-            $ticket->category_id = $request->category_id;
-            $ticket->save();
+            if ($user->hasVerifiedEmail()) {
+                $ticket = new Ticket();
+                $ticket->user_id = $user->id;
+                $ticket->title = $request->title;
+                $ticket->category_id = $request->category_id;
+                $ticket->save();
 
-            //create ticket content
-            $ticket_content = TicketContentController::add_content($ticket->id, $user->id, $request->content_type, $request->text);
+                //create ticket content
+                $ticket_content = TicketContentController::add_content($ticket->id, $user->id, $request->content_type, $request->text);
 
-            // if ticket is created
-            if ($ticket && $ticket_content) {
-                return response()->json(['message' => 'Ticket created ID : '.$ticket->id.''], 201);
+                // if ticket is created
+                if ($ticket && $ticket_content) {
+                    return response()->json(['message' => 'Ticket created ID : ' . $ticket->id . ''], 201);
+                } else {
+                    return response()->json(['error' => 'Ticket not created'], 500);
+                }
             } else {
-                return response()->json(['error' => 'Ticket not created'], 500);
+                return response()->json(['error' => 'Verify email address'], 403);
             }
         } else {
             return response()->json(['error' => 'Verfiy api token'], 403);
@@ -97,27 +105,31 @@ class TicketController extends Controller
         $api_token = $request->header('Authorization');
         $user = UserController::verify_token($api_token);
         if ($user) {
-            //get ticket by id
-            $ticket = Ticket::find($request->id);
-            if ($ticket) {
-                //verify user has permission to read ticket or is owner of ticket
-                if ($user->hasPerm('read-ticket') || $user->id == $ticket->user_id) {
-                    //get assigned ticket by ticket id
-                    $assigned_ticket = Ticket::where('id', $ticket->id)->first()->assignedUser()->get()->first();
-                    if ($assigned_ticket) {
-                        //if ticket is assigned then set assigned to true and get user id
-                        $ticket->assigned = true;
-                        $ticket->assigned_user = $assigned_ticket->id;
-                    } else {
-                        $ticket->assigned = false;
-                    }
+            if ($user->hasVerifiedEmail()) {
+                //get ticket by id
+                $ticket = Ticket::find((int) $request->id);
+                if ($ticket) {
+                    //verify user has permission to read ticket or is owner of ticket
+                    if ($user->hasPerm('read-ticket') || $user->id == $ticket->user_id) {
+                        //get assigned ticket by ticket id
+                        $assigned_ticket = Ticket::where('id', $ticket->id)->first()->assignedUser()->get()->first();
+                        if ($assigned_ticket) {
+                            //if ticket is assigned then set assigned to true and get user id
+                            $ticket->assigned = true;
+                            $ticket->assigned_user = $assigned_ticket->id;
+                        } else {
+                            $ticket->assigned = false;
+                        }
 
-                    return response()->json($ticket, 200);
+                        return response()->json($ticket, 200);
+                    } else {
+                        return response()->json(['error' => 'You have not the right permission to read the ticket'], 404);
+                    }
                 } else {
-                    return response()->json(['error' => 'You have not the right permission to read the ticket'], 404);
+                    return response()->json(['error' => 'Ticket not found'], 404);
                 }
             } else {
-                return response()->json(['error' => 'Ticket not found'], 404);
+                return response()->json(['error' => 'Verify email address'], 403);
             }
         } else {
             return response()->json(['error' => 'Verfiy api token'], 403);
@@ -138,19 +150,23 @@ class TicketController extends Controller
         $api_token = $request->header('Authorization');
         $user = UserController::verify_token($api_token);
         if ($user) {
-            //get ticket by id
-            $ticket = Ticket::find($request->id);
-            if ($ticket) {
-                //verify if user is assigned to ticket or is admin
-                if ($user->id === $ticket->assignedUser || $user->hasPerm('update-ticket')) {
-                    if (Ticket::update_header($request->id, $request->title, $request->category_id, $request->rating, $request->category, $request->status)) {
-                        return response()->json(['message' => 'Ticket updated'], 200);
-                    } else {
-                        return response()->json(['error' => 'Ticket not updated'], 500);
+            if ($user->hasVerifiedEmail()) {
+                //get ticket by id
+                $ticket = Ticket::find($request->id);
+                if ($ticket) {
+                    //verify if user is assigned to ticket or is admin
+                    if ($user->id === $ticket->assignedUser || $user->hasPerm('update-ticket')) {
+                        if (Ticket::update_header($request->id, $request->title, $request->category_id, $request->rating, $request->category, $request->status)) {
+                            return response()->json(['message' => 'Ticket updated'], 200);
+                        } else {
+                            return response()->json(['error' => 'Ticket not updated'], 500);
+                        }
                     }
+                } else {
+                    return response()->json(['error' => 'You are not assigned or owner of the ticket'], 403);
                 }
             } else {
-                return response()->json(['error' => 'You are not assigned or owner of the ticket'], 403);
+                return response()->json(['error' => 'Verify email address'], 403);
             }
         } else {
             return response()->json(['error' => 'Verfiy api token'], 403);
@@ -170,21 +186,27 @@ class TicketController extends Controller
         $api_token = $request->header('Authorization');
         $user = UserController::verify_token($api_token);
         if ($user) {
-            //get ticket by id
-            $ticket = Ticket::find($request->id);
-            if ($ticket) {
-                //verify if the user is assigned to ticket or has permission to delete ticket
-                if ($user->id === $ticket->assignedUser || $user->hasPerm('delete-ticket')) {
-                    //delete ticket
-                    if (Ticket::find($request->id)->delete()) {
-                        return response()->json(['message' => 'Ticket deleted'], 200);
+            if ($user->hasVerifiedEmail()) {
+                //get ticket by id
+                $ticket = Ticket::find($request->id);
+                if ($ticket) {
+                    //verify if the user is assigned to ticket or has permission to delete ticket
+                    if ($user->id === $ticket->assignedUser || $user->hasPerm('delete-ticket')) {
+                        //delete ticket
+                        if (Ticket::find($request->id)->delete()) {
+                            return response()->json(['message' => 'Ticket deleted'], 200);
+                        } else {
+                            return response()->json(['error' => 'Ticket not deleted'], 500);
+                        }
                     } else {
-                        return response()->json(['error' => 'Ticket not deleted'], 500);
+                        return response()->json(['error' => 'You are not assigned or owner of the ticket'], 403);
                     }
-                } else {
-                    return response()->json(['error' => 'You are not assigned or owner of the ticket'], 403);
                 }
+            } else {
+                return response()->json(['error' => 'Verify email address'], 403);
             }
+        } else {
+            return response()->json(['error' => 'Verfiy api token'], 403);
         }
     }
 }
